+++ packages/zosfiles/__tests__/__unit__/methods/create/Create.unit.test.ts
@@ -251,6 +251,133 @@ describe("Create data set", () => {
             );
         });
 
+        it("should be able to create a variable block sequential data set using a block size that is too small", async () => {
+            const custOptions = {
+                dsorg: "PS",
+                alcunit: "CYL",
+                primary: 20,
+                secondary: 1,
+                recfm: "VB",
+                blksize: 100,
+                lrecl: 1000
+            };
+            const response = await Create.dataSet(dummySession, CreateDataSetTypeEnum.DATA_SET_SEQUENTIAL, dataSetName, custOptions);
+
+            expect(response.success).toBe(true);
+            expect(response.commandResponse).toContain("created successfully");
+            expect(mySpy).toHaveBeenCalledWith(
+                dummySession,
+                endpoint,
+                [ZosmfHeaders.ACCEPT_ENCODING],
+                JSON.stringify({
+                    ...{
+                        alcunit: "CYL",
+                        dsorg: "PS",
+                        primary: 20,
+                        recfm: "VB",
+                        blksize: 1004,
+                        lrecl: 1000,
+                        secondary: 1
+                    }
+                })
+            );
+        });
+
+        it("should be able to create a fixed block sequential data set using a block size that is too small", async () => {
+            const custOptions = {
+                dsorg: "PS",
+                alcunit: "CYL",
+                primary: 20,
+                secondary: 1,
+                recfm: "FB",
+                blksize: 100,
+                lrecl: 1000
+            };
+            const response = await Create.dataSet(dummySession, CreateDataSetTypeEnum.DATA_SET_SEQUENTIAL, dataSetName, custOptions);
+
+            expect(response.success).toBe(true);
+            expect(response.commandResponse).toContain("created successfully");
+            expect(mySpy).toHaveBeenCalledWith(
+                dummySession,
+                endpoint,
+                [ZosmfHeaders.ACCEPT_ENCODING],
+                JSON.stringify({
+                    ...{
+                        alcunit: "CYL",
+                        dsorg: "PS",
+                        primary: 20,
+                        recfm: "FB",
+                        blksize: 1000,
+                        lrecl: 1000,
+                        secondary: 1
+                    }
+                })
+            );
+        });
+
+        it("should be able to create a variable sequential data set using a block size that is too small", async () => {
+            const custOptions = {
+                dsorg: "PS",
+                alcunit: "CYL",
+                primary: 20,
+                secondary: 1,
+                recfm: "V",
+                blksize: 100,
+                lrecl: 1000
+            };
+            const response = await Create.dataSet(dummySession, CreateDataSetTypeEnum.DATA_SET_SEQUENTIAL, dataSetName, custOptions);
+
+            expect(response.success).toBe(true);
+            expect(response.commandResponse).toContain("created successfully");
+            expect(mySpy).toHaveBeenCalledWith(
+                dummySession,
+                endpoint,
+                [ZosmfHeaders.ACCEPT_ENCODING],
+                JSON.stringify({
+                    ...{
+                        alcunit: "CYL",
+                        dsorg: "PS",
+                        primary: 20,
+                        recfm: "V",
+                        blksize: 1004,
+                        lrecl: 1000,
+                        secondary: 1
+                    }
+                })
+            );
+        });
+
+        it("should be able to create a fixed sequential data set using a block size that is too small without specifying the alcunit", async () => {
+            const custOptions = {
+                dsorg: "PS",
+                alcunit: "CYL",
+                primary: 20,
+                secondary: 1,
+                blksize: 100,
+                lrecl: 1000
+            };
+            const response = await Create.dataSet(dummySession, CreateDataSetTypeEnum.DATA_SET_SEQUENTIAL, dataSetName, custOptions);
+
+            expect(response.success).toBe(true);
+            expect(response.commandResponse).toContain("created successfully");
+            expect(mySpy).toHaveBeenCalledWith(
+                dummySession,
+                endpoint,
+                [ZosmfHeaders.ACCEPT_ENCODING],
+                JSON.stringify({
+                    ...{
+                        alcunit: "CYL",
+                        dsorg: "PS",
+                        primary: 20,
+                        recfm: "FB",
+                        blksize: 1000,
+                        lrecl: 1000,
+                        secondary: 1
+                    }
+                })
+            );
+        });
+
         it("should be able to create a sequential data set using the primary allocation and default the secondary allocation", async () => {
             const custOptions = {
                 dsorg: "PS",
@@ -307,7 +434,7 @@ describe("Create data set", () => {
                 size: "3TRK",
                 secondary: 2,
                 recfm: "VB",
-                blksize: 4096,
+                blksize: 4100,
                 lrecl: 4096,
                 dirblk: 5
             };
@@ -419,7 +546,7 @@ describe("Create data set", () => {
                 size: "TRK",
                 secondary: 2,
                 recfm: "VB",
-                blksize: 4096,
+                blksize: 4100,
                 lrecl: 4096,
                 dirblk: 5
             };